Examples: fix deployment to work without config affects issues #119 #85
