+++ entity-services-examples/gen/ml-modules/ext/Race-0.0.1.xqy
@@ -20,7 +20,7 @@ xquery version "1.0-ml";
 (: database of your application, and check it into your source control system.      :)
 (:                                                                                  :)
 (: Modification History:                                                            :)
-(: Generated at timestamp: 2016-08-22T10:17:07.062645-07:00                         :)
+(: Generated at timestamp: 2016-08-25T12:06:31.76438-07:00                          :)
 (:   Persisted by AUTHOR                                                            :)
 (:   Date: DATE                                                                     :)
 module namespace race
