+++ entity-services-examples/gen/ml-modules/ext/Race-0.0.2.xqy
@@ -20,7 +20,7 @@ xquery version "1.0-ml";
 (: database of your application, and check it into your source control system.      :)
 (:                                                                                  :)
 (: Modification History:                                                            :)
-(: Generated at timestamp: 2016-08-22T10:17:07.062645-07:00                         :)
+(: Generated at timestamp: 2016-08-25T12:06:31.76438-07:00                          :)
 (:   Persisted by AUTHOR                                                            :)
 (:   Date: DATE                                                                     :)
 module namespace race 
@@ -87,6 +87,7 @@ declare function race:extract-instance-Race(
 (: The output of this function should structurally match the output of              :)
 (: es:model-get-test-instances($model)                                              :)
 (:                                                                                  :)
+ =>es:optional('id',                     sem:iri($source-node/Race/id))
  =>   map:with('name',                   xs:string($source-node/Race/name))
  (: The following property assigment comes from an external reference.               :)
  (: Its generated value probably requires developer attention.                       :)
@@ -144,7 +145,7 @@ declare function race:extract-instance-Run(
 (: The output of this function should structurally match the output of              :)
 (: es:model-get-test-instances($model)                                              :)
 (:                                                                                  :)
- =>   map:with('id',                     xs:string($source-node/Run/id))
+ =>   map:with('id',                     sem:iri($source-node/Run/id))
  =>   map:with('date',                   xs:date($source-node/Run/date))
  =>   map:with('distance',               xs:decimal($source-node/Run/distance))
  =>es:optional('distanceLabel',          xs:string($source-node/Run/distanceLabel))
@@ -199,6 +200,7 @@ declare function race:extract-instance-Runner(
 (: The output of this function should structurally match the output of              :)
 (: es:model-get-test-instances($model)                                              :)
 (:                                                                                  :)
+ =>es:optional('id',                     sem:iri($source-node/Runner/id))
  =>   map:with('name',                   xs:string($source-node/Runner/name))
  =>   map:with('age',                    xs:decimal($source-node/Runner/age))
  =>es:optional('gender',                 xs:string($source-node/Runner/gender))
