+++ entity-services-examples/gen/ml-modules/ext/translator.xqy
@@ -13,7 +13,7 @@ import module namespace es = "http://marklogic.com/entity-services"
 (: Race, version 0.0.1                                                              :)
 (:                                                                                  :)
 (: Modification History:                                                            :)
-(: Generated at timestamp: 2016-08-22T10:17:07.062645-07:00                         :)
+(: Generated at timestamp: 2016-08-25T12:06:31.76438-07:00                          :)
 (:   Persisted by AUTHOR                                                            :)
 (:   Date: DATE                                                                     :)
 
@@ -33,6 +33,8 @@ declare function race-from-race:convert-instance-Race(
 (: The following line identifies the type of this instance.  Do not change it.      :)
  =>map:with('$type', 'Race')
 (: The following lines are generated from the 'Race' entity type.                   :)
+ (: The following property was missing from the source type                          :)
+=>es:optional('id',                     sem:iri( () ))
  =>   map:with('name',                   xs:string($source-node/Race/name))
  (: The following property was missing from the source type                          :)
 =>es:optional('raceCategory',           xs:string( () ))
@@ -57,7 +59,7 @@ declare function race-from-race:convert-instance-Run(
 (: The following line identifies the type of this instance.  Do not change it.      :)
  =>map:with('$type', 'Run')
 (: The following lines are generated from the 'Run' entity type.                    :)
- =>   map:with('id',                     xs:string($source-node/Run/id))
+ =>   map:with('id',                     sem:iri($source-node/Run/id))
  =>   map:with('date',                   xs:date($source-node/Run/date))
  =>   map:with('distance',               xs:decimal($source-node/Run/distance))
  =>es:optional('distanceLabel',          xs:string($source-node/Run/distanceLabel))
@@ -81,6 +83,8 @@ declare function race-from-race:convert-instance-Runner(
 (: The following line identifies the type of this instance.  Do not change it.      :)
  =>map:with('$type', 'Runner')
 (: The following lines are generated from the 'Runner' entity type.                 :)
+ (: The following property was missing from the source type                          :)
+=>es:optional('id',                     sem:iri( () ))
  =>   map:with('name',                   xs:string($source-node/Runner/name))
  =>   map:with('age',                    xs:decimal($source-node/Runner/age))
  =>es:optional('gender',                 xs:string($source-node/Runner/gender))
