+++ entity-services-examples/src/main/java/com/marklogic/entityservices/examples/ExamplesLoader.java
@@ -16,124 +16,23 @@
 package com.marklogic.entityservices.examples;
 
 import java.io.IOException;
+import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.concurrent.TimeUnit;
 
-import com.marklogic.client.query.StructuredQueryDefinition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.marklogic.client.document.ServerTransform;
-import com.marklogic.client.query.StructuredQueryBuilder;
-import com.marklogic.datamovement.ApplyTransformListener;
-import com.marklogic.datamovement.JobTicket;
-import com.marklogic.datamovement.QueryHostBatcher;
 
 /**
  * Runs the load methods for entity services, rdf, and json instances.
  */
-public class ExamplesLoader extends ExamplesBase {
+public class ExamplesLoader {
 
     private static Logger logger = LoggerFactory.getLogger(ExamplesLoader.class);
 
-    public ExamplesLoader() throws IOException {
-        super();
-    }
-
-    public Thread modelsLoad() {
-        Runnable task = () -> {
-            try {
-                importJSON(Paths.get(props.getProperty("projectDir") + "/data/models"),
-                        "http://marklogic.com/entity-services/models");
-            } catch (IOException e) {
-                logger.error("IOException thrown by loader.");
-            }
-        };
-        task.run();
-        return new Thread(task);
-    }
-
-    public Thread instanceLoad() {
-        Runnable task = () -> {
-            try {
-                importJSON(Paths.get(props.getProperty("projectDir") + "/data/race-data"), "raw");
-            } catch (IOException e) {
-                logger.error("IOException thrown by loader.");
-            }
-        };
-        task.run();
-        return new Thread(task);
-    }
-
-    public Thread rdfLoad() {
-        Runnable task = () -> {
-            importRDF(Paths.get(props.getProperty("projectDir") + "/data/third-party/rdf"), "reference");
-        };
-        task.run();
-        return new Thread(task);
-    }
-
-    public void harmonize() throws InterruptedException {
-        StructuredQueryBuilder qb = new StructuredQueryBuilder();
-        qb.collection("raw");
-        ServerTransform ingester = new ServerTransform("ingester");
-        ApplyTransformListener listener = new ApplyTransformListener().withTransform(ingester)
-                .withApplyResult(ApplyTransformListener.ApplyResult.IGNORE).onSuccess((dbClient, inPlaceBatch) -> {
-                    logger.debug("Batch transform SUCCESS");
-                }).onBatchFailure((dbClient, inPlaceBatch, throwable) -> {
-                    // logger.warn("FAILURE on batch:" + inPlaceBatch.toString()
-                    // + "\n", throwable);
-                    // throwable.printStackTrace();
-                    System.err.println(throwable.getMessage());
-                    System.err.print(String.join("\n", inPlaceBatch.getItems()) + "\n");
-                });
-
-        QueryHostBatcher queryHostBatcher = moveMgr.newQueryHostBatcher(qb.build()).withBatchSize(100)
-                .withThreadCount(5).onUrisReady(listener).onQueryFailure((client3, exception) -> {
-                    logger.error("Query error");
-                });
-
-        JobTicket ticket = moveMgr.startJob(queryHostBatcher);
-        queryHostBatcher.awaitTermination(Long.MAX_VALUE, TimeUnit.DAYS);
-        moveMgr.stopJob(ticket);
-    }
-
-    public void secondSourceHarmonize() throws InterruptedException {
-        StructuredQueryBuilder qb = new StructuredQueryBuilder();
-        StructuredQueryDefinition qdef = qb.collection("csv");
-        ServerTransform ingester = new ServerTransform("ingester-angel-island");
-        ApplyTransformListener listener = new ApplyTransformListener().withTransform(ingester)
-                .withApplyResult(ApplyTransformListener.ApplyResult.IGNORE).onSuccess((dbClient, inPlaceBatch) -> {
-                    logger.debug("batch transform SUCCESS");
-                }).onBatchFailure((dbClient, inPlaceBatch, throwable) -> {
-                    logger.error("FAILURE on batch:" + inPlaceBatch.toString() + "\n", throwable);
-                    //System.err.println(throwable.getMessage());
-                    //System.err.print(String.join("\n", inPlaceBatch.getItems()) + "\n");
-                });
-
-        QueryHostBatcher queryHostBatcher = moveMgr //
-                .newQueryHostBatcher(qdef) //
-                .withBatchSize(10) //
-                .withThreadCount(5) //
-                .onUrisReady(listener) //
-                .onQueryFailure((client3, exception) -> {
-                    logger.error("Query error");
-                });
-
-        JobTicket ticket = moveMgr.startJob(queryHostBatcher);
-        queryHostBatcher.awaitTermination(Long.MAX_VALUE, TimeUnit.DAYS);
-        moveMgr.stopJob(ticket);
-    }
-
-    private void loadAsIs() {
-        modelsLoad().start();
-        instanceLoad().start();
-        rdfLoad().start();
-    }
-
     public static void main(String[] args) throws IOException, InterruptedException {
 
-        ExamplesLoader loader = new ExamplesLoader();
+        AsIsLoader loader = new AsIsLoader();
         loader.loadAsIs();
 
         CSVLoader integrator = new CSVLoader();
@@ -142,8 +41,9 @@ public class ExamplesLoader extends ExamplesBase {
         logger.info("Pausing to let load proceed before harmonize");
         Thread.sleep(20000);
         logger.info("Starting harmonize");
-        loader.harmonize();
-        loader.secondSourceHarmonize();
+        Harmonizer harmonizer = new Harmonizer();
+        harmonizer.harmonize();
+        harmonizer.secondSourceHarmonize();
 
         CodeGenerator generator = new CodeGenerator();
         generator.generate();
