--- entity-services-examples/src/main/resources/application.properties.CHANGEME
@@ -1,13 +1,0 @@
-# Change this to reflect where you’ve checked out the code,
-# specifically the entity-services-examples subdirectory
-projectDir=/Users/jmakeig/Workspaces/entity-services/entity-services-examples
-
-# These values typically do not need to be changed
-# They align with the users created with mlDeploy
-mlHost=localhost
-mlUsername=whitney
-mlPassword=********
-mlRestPort=8203
-mlAdminUsername=adam
-mlAdminPassword=********
-
